ProactiveAppinstallation c# code
