+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31112.23
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProactiveBot", "Proactive_Installation _Sample\ProactiveBot.csproj", "{887C82B6-0C97-4B4C-AE7E-F9CDD98A3BD1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProactiveAppInstallation", "ProactiveAppInstallation\ProactiveAppInstallation.csproj", "{887C82B6-0C97-4B4C-AE7E-F9CDD98A3BD1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
