+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/AdapterWithErrorHandler.cs
