+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/Models/AppModel.cs
@@ -1,0 +1,10 @@
+﻿namespace ProactiveBot.Models
+{
+    public class InstallationCounts
+    {
+        public int Existing { get; set; }
+
+        public int New { get; set; }
+    }
+
+}
