+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/ProactiveAppInstallation.csproj
@@ -18,7 +18,7 @@
     <PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="4.12.1" />
     <PackageReference Include="Microsoft.Bot.Connector" Version="4.13.2" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Graph" Version="3.29.0" />
+    <PackageReference Include="Microsoft.Graph" Version="3.34.0" />
     <PackageReference Include="Microsoft.Graph.Auth" Version="1.0.0-preview.7" />
     <PackageReference Include="Microsoft.Identity.Client" Version="4.30.1" />
     <PackageReference Include="Microsoft.OData.Client" Version="7.8.3" />
