+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/Program.cs
