+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/Properties/launchSettings.json
