+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/Startup.cs
