+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/TeamsAppManifest/color.png
