+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/TeamsAppManifest/outline.png
