+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/appsettings.Development.json
