+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/appsettings.json
@@ -1,0 +1,6 @@
+﻿
+{
+  "MicrosoftAppId": "<<YOUR-MICROSOFT-APP-ID>>",
+  "MicrosoftAppPassword": "<<YOUR-MICROSOFT-APP-PASSWORD>>",
+  "AppCatalogTeamAppId": "<<YOUR--APPCATALOG-TEAM-ID>>"
+}
