+++ samples/graph-proactive-installation/csharp/ProactiveAppInstallation/wwwroot/default.htm
