--- samples/graph-proactive-installation/csharp/Proactive_Installation _Sample/Resources/Exist_AdaptiveCard.json
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-{
-  "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
-  "version": "1.0",
-  "type": "AdaptiveCard",
-  "body": [
-    {
-      "type": "TextBlock",
-      "text": "This App already exists in your personal Scope",
-      "weight": "bolder",
-      "isSubtle": false
-    }
-  ]
-}
