--- samples/graph-proactive-installation/csharp/Proactive_Installation _Sample/Resources/New_AdaptiveCard.json
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-{
-  "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
-  "version": "1.0",
-  "type": "AdaptiveCard",
-  "body": [
-    {
-      "type": "TextBlock",
-      "text": "Welcome to Proactive bot Sample",
-      "weight": "bolder",
-      "isSubtle": false
-    }
-  ]
-}
