--- samples/graph-proactive-installation/csharp/Proactive_Installation _Sample/Views/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "_Layout";
-}
