--- samples/graph-proactive-installation/csharp/Proactive_Installation _Sample/appsettings.json
@@ -1,5 +1,0 @@
-﻿{
-  "MicrosoftAppId": "<<YOUR-MICROSOFT-APP-ID>>",
-  "MicrosoftAppSecret": "<<YOUR-MICROSOFT-APP-PASSWORD>>",
-  "AppCatalogTeamAppId": "<<YOUR-MICROSOFT-TEAM-APP-ID>>"
-}
